feat: add support for API Tokens, ISO Images, VPC Networks and core SDK updates
New features:
- Added services for API Tokens, ISO Images, VPC Networks, Languages, Storage Tags, and Usage.
- Updated Client to register new services.

Core updates:
- Updated request handling to support 200 OK on DELETE.
- Added StopWithRequest to VirtualServersService.

Maintenance:
- Cosmetic documentation updates in generated files.

+++ additionalDisks.go
@@ -11,7 +11,7 @@ type Disk struct {
 	IsPrimary  bool    `json:"is_primary"`
 	Name       string  `json:"name"`
 	Size       int     `json:"size"`
-	ActualSize int     `json:"actual_size"`
+	ActualSize float64 `json:"actual_size"`
 	Path       string  `json:"path"`
 	FullPath   string  `json:"full_path"`
 	Storage    Storage `json:"storage"`
